심준보 : 몬스터 스크립트 분할, 몬스터 애니메이션 수정, 몬스터 프리팹 완성(근접예티, 근접오크, 근접정령, 원거리정령), 몬스터 오브젝트 풀 스크립트 작성 중, Mod : Monster Script divided, Animation Fixed, Prefabs Ready, Monster ObjPool Script added

+++ ProjectM/Assets/2. Test/4. SJB/Scripts/Monster_CSV/MonsterData.cs
@@ -4,8 +4,8 @@ public class MonsterData
     public virtual int MonsterID { get; protected set; }
     public virtual string MonsterDescription { get; protected set; }
     public virtual int MonsterType { get; protected set; }
-    public virtual int MonsterAttackRange { get; protected set; }
-    public virtual int MonsterMoveSpeed { get; protected set; }
+    public virtual float MonsterAttackRange { get; protected set; }
+    public virtual float MonsterMoveSpeed { get; protected set; }
     public virtual int MonsterHP { get; protected set; }
     public virtual int MonsterDamage { get; protected set; }
     public virtual int MonsterRunSpeed { get; protected set; }
